Add helper to do GitHub API requests
